backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pillow
opencv-python-headless
tensorflow-cpu
torch
torchvision
pdfplumber
pytesseract
google-generativeai
cloudinary
redis
httpx
python-dotenv
alembic
slowapi
python-magic
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
openai==1.3.0
python-dotenv==1.0.0
slowapi==0.1.9
redis==5.0.1
